优化childController生命周期调用方法

+++ CMPageTitleView/CMPageTitleView/Demo/CMViewController.m
@@ -1673,14 +1673,14 @@ isPhoneX = [[UIApplication sharedApplication] delegate].window.safeAreaInsets.bo
 
 - (void)cm_pageTitleViewSelectedWithIndex:(NSInteger)index Repeat:(BOOL)repeat {
     
-    NSLog(@"第%ld个标题被选中；且%@是重复选中",index,repeat?@"":@"不");
+    NSLog(@"第%ld个标题被选中；且%@是重复选中",(long)index,repeat?@"":@"不");
     
 }
 
 - (void)cm_pageTitleViewClickWithIndex:(NSInteger)index Repeat:(BOOL)repeat {
     
     
-    NSLog(@"第%ld个标题被点击；且%@是重复点击",index,repeat?@"":@"不");
+    NSLog(@"第%ld个标题被点击；且%@是重复点击",(long)index,repeat?@"":@"不");
     
 }
 
@@ -1688,9 +1688,10 @@ isPhoneX = [[UIApplication sharedApplication] delegate].window.safeAreaInsets.bo
 - (void)cm_pageTitleViewScrollToIndex:(NSInteger)index {
     
     
-    NSLog(@"当前滚动到第%ld个标题",index);
+    NSLog(@"当前滚动到第%ld个标题",(long)index);
     
 }
+
 
 
 @end
